angle

+++ RobotGoFast/RobotState.h
@@ -7,6 +7,6 @@ enum RobotState { IDLE, RUNNING, ERROR, COMPLETE };
 
 const double IR_SPACING = 4.6625;
 
-void printDouble(double val, unsigned int precision);
+void printDouble(float val, unsigned int precision);
 
 #endif // ROBOTSTATE_H
